+++ android/src/org/coolreader/plugins/litres/LitresPlugin.java
@@ -41,17 +41,19 @@ import org.coolreader.plugins.OnlineStorePlugin;
 import org.coolreader.plugins.litres.LitresConnection.LitresAuthInfo;
 import org.coolreader.plugins.litres.LitresConnection.ResultHandler;
 
-import android.app.Activity;
+import android.content.Context;
 import android.content.SharedPreferences;
 
 public class LitresPlugin implements OnlineStorePlugin {
 
 	public static final String PACKAGE_NAME = "org.coolreader.plugins.litres";
-	private Activity activity;
+	private final Context applicationContext;
 	
 	private final LitresConnection connection;
-	public LitresPlugin(Activity activity, SharedPreferences preferences) {
-		this.activity = activity;
+	public LitresPlugin(
+			Context applicationContext,
+			SharedPreferences preferences) {
+		this.applicationContext = applicationContext;
 		connection = LitresConnection.create(preferences);
 	}
 	
@@ -70,12 +72,14 @@ public class LitresPlugin implements OnlineStorePlugin {
 
 	@Override
 	public String getDescription() {
-		return activity.getString(org.coolreader.R.string.online_store_plugin_description_litres);
+		return applicationContext.getString(
+				org.coolreader.R.string.online_store_plugin_description_litres);
 	}
 
 	@Override
 	public String getName() {
-		return activity.getString(org.coolreader.R.string.online_store_plugin_name_litres);
+		return applicationContext.getString(
+				org.coolreader.R.string.online_store_plugin_name_litres);
 	}
 
 	@Override